rename:constants/ 디렉토리 위치 변경
* src/components/ -> src/
